cleanup: separated reminder service file
